+++ crates/common/src/lib.rs
@@ -1,4 +1,4 @@
+pub mod capabilities_response;
 pub mod client;
 pub mod config;
-pub mod config_file;
-pub mod schema;
+pub mod schema_response;